errors should be uint16_t

+++ include/quicly.h
@@ -374,14 +374,14 @@ struct st_quicly_stream_t {
          */
         struct {
             quicly_sender_state_t sender_state;
-            uint32_t reason;
+            uint16_t reason;
         } stop_sending;
         /**
          * rst_stream
          */
         struct {
             quicly_sender_state_t sender_state;
-            uint32_t reason;
+            uint16_t reason;
         } rst;
         /**
          * sends receive window updates to peer
@@ -539,7 +539,7 @@ static int quicly_stream_is_closable(quicly_stream_t *stream);
 /**
  *
  */
-void quicly_reset_stream(quicly_stream_t *stream, unsigned direction, uint32_t reason);
+void quicly_reset_stream(quicly_stream_t *stream, unsigned direction, uint16_t reason);
 /**
  *
  */
